lamnt98 commit chuyển float thành double 22/07

+++ project/src/main/java/com/example/webDemo3/service/impl/manageSchoolRankImpl/DownloadRankWeekServiceImpl.java
@@ -89,7 +89,7 @@ public class DownloadRankWeekServiceImpl implements DownloadRankWeekService {
      * @param input
      * @return
      */
-    private double round(Float input) {
+    private double round(Double input) {
         return (double) Math.round(input * 100) / 100;
     }
 }
